+++ internal/appinfo/appinfo.go
@@ -8,7 +8,7 @@ import (
 
 const (
 	Name        = "hellogrok"
-	Version     = "0.1.16"
+	Version     = "0.1.17"
 	LogFileName = "hellogrok.log"
 )